+++ src/components/researchPortal/ResearchExpressionPlot.vue
@@ -6,12 +6,7 @@
         <div v-if="!showPlot">
             <p>No datasets meet minimum sample count.</p>
         </div>
-        <button
-            class="btn btn-primary btn-sm download-chart"
-            @click="downloadChart()"
-        >
-            Download plot (.svg)
-        </button>
+        <svg-download chartClass="svg-chart"></svg-download>
     </div>
 </template>
 <script>
@@ -19,6 +14,7 @@ import Vue from "vue";
 import * as d3 from "d3";
 import colors from "@/utils/colors";
 import Formatters from "@/utils/formatters";
+import SvgDownload from "../SvgDownload.vue";
 export default Vue.component("ResearchExpressionPlot", {
     props: ["plotData", "highlightedDataset"],
     data() {
@@ -309,26 +305,6 @@ export default Vue.component("ResearchExpressionPlot", {
                 .style("width", 50)
                 .on("mouseover", (d) => this.hoverViolin(d.key));
         },
-        downloadChart() {
-            // Serialize the SVG to a string
-            const svgString = new XMLSerializer().serializeToString(
-                d3.select("svg.svg-chart").node()
-            );
-
-            // Create a data URL
-            const blob = new Blob([svgString], {
-                type: "image/svg+xml;charset=utf-8",
-            });
-            const url = URL.createObjectURL(blob);
-
-            // Create a link element and programmatically click it to start the download
-            const link = document.createElement("a");
-            link.href = url;
-            link.download = "chart.svg";
-            document.body.appendChild(link);
-            link.click();
-            document.body.removeChild(link);
-        },
         getBottomMargin(data) {
             let longestLabel = data
                 .map((item) => item.keyField.length)
@@ -514,9 +490,5 @@ div {
 }
 .tooltip span {
     display: block;
-}
-button.download-chart {
-    float: right;
-    margin-bottom: 25px;
 }
 </style>
